- adds announcement logic back to base activity
